pyproject.toml
[project]
name = "talker-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "celery>=5.6.0",
    "email-validator>=2.3.0",
    "fastapi>=0.125.0",
    "google-auth>=2.45.0",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "langchain>=1.2.0",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pgvector>=0.4.2",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.5",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.38.0",
]
